Implements the initial workflow (#1)
TL;DR
-----

Fleshes out a manually run workflow for creating a demo

Details
-------

Provides a complete implementation of a workflow run using the
GitHub workflow dispatch option. The workflow is used to build
SlackerNews and release it to a GitHub account/org and Replicated
vendor team other than the main Slackernews ones.

To use it, run the action from the "Actions" tab and specify the
following:

1. The branch on the Slackernews repo you want to build from
2. The namespace on the `ghcr.io` registry to push images to.
3. Your application slug
4. The version for your release to `Unstable`
5. The domain to use for the proxy image registry

The workflow is the same as the release workflow for Slackerenws.
In the end you'll have an images pushed to your own namespace on 
the GitHub registry and a release promoted to the `Unstable` 
channel in your own vendor tea. That team will pull images through 
the proxy using the domain you specify.